requirements.txt
pandas
numpy
scikit-learn
xgboost
shap
streamlit
matplotlib
seaborn
plotly
mysql-connector-python
jupyter
openpyxl
pillow